Jarvis_AI/requirements.txt
speechrecognition
pyttsx4
requests
google-generativeai
pyaudio